♻️ Moved reused function to seperate file
